pyproject.toml
[tool.poetry]
name = "workspace"
version = "0.1.0"
description = ""
authors = ["Lucas <lucasmiguelmac@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "3.2"
psycopg2-binary = "^2.9.3"
django-cors-headers = "^3.11.0"
djangorestframework = "^3.13.1"
six = "^1.16.0"
idna = "^3.3"
django-phonenumber-field = "^6.1.0"
phonenumbers = "^8.12.49"
babel = "^2.10.1"
django_rest_framework = "^0.1.0"
requests-mock = "^1.9.3"
pytest-mock = "^3.7.0"

[tool.poetry.dev-dependencies]
django-extensions = "^3.1.5"
pre-commit = "^2.19.0"
autopep8 = "^1.6.0"
isort = "^5.10.1"
pytest = "^7.1.2"
pytest-django = "^4.5.2"
pytest-cov = "^3.0.0"
model_bakery = "^1.5.0"
django-mock-queries = "^2.1.7"
ipython = "^8.4.0"

[tool.pytest]
addopts="-vv"
python_files = "test_*.py *_tests.py"
DJANGO_SETTINGS_MODULE = "project.settings.base"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

